admin/requirements.txt
django
djangorestframework
mysqlclient
django-mysql
django-cors-headers
pika
